fix typo

+++ json/cJSON.h
@@ -106,7 +106,7 @@ extern cJSON *cJSON_CreateObject(void);
 /* These utilities create an Array of count items. */
  // taka add
 extern cJSON *cJSON_CreateUInt8Array(const uint8_t *numbers,int count);
-extern cJSON *cJSON_CreateUInt16Array(const uint8_t *numbers,int count);
+extern cJSON *cJSON_CreateUInt16Array(const uint16_t *numbers,int count);
 
 extern cJSON *cJSON_CreateIntArray(const int *numbers,int count);
 extern cJSON *cJSON_CreateFloatArray(const float *numbers,int count);
